+++ apps/testbed/src/auth/auth-client.ts
@@ -1,6 +1,6 @@
-import { cfxClient } from "@homestead/ba-cfx/client";
-import { steamClient } from "@homestead/ba-steam/client";
-import { tebexClient } from "@homestead/ba-tebex/client";
+import { cfxClient } from "@homestead-systems/ba-cfx/client";
+import { steamClient } from "@homestead-systems/ba-steam/client";
+import { tebexClient } from "@homestead-systems/ba-tebex/client";
 import { createAuthClient } from "better-auth/react";
 
 export const authClient = createAuthClient({