+++ packages/cfx/CHANGELOG.md
@@ -1,3 +1,9 @@
 # Changelog
 
+## 0.1.0
+
+### Minor Changes
+
+- Initial public release of the Cfx.re and Steam Better Auth plugins.
+
 This file is maintained by Changesets after the package's first release.